requirements.txt
astroid==1.6.1
boto3==1.7.4
botocore==1.10.4
colorama==0.3.9
Django==1.11.3
django-coverage==1.2.4
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docutils==0.14
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.1
PyJWT==1.6.1
pylint==1.7.2
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
s3transfer==0.1.13
six==1.11.0
wrapt==1.10.11
